+++ BreastCancer/TasksAndSteps/Common/APHCommonTaskSummaryViewController.m
@@ -65,7 +65,7 @@ static NSString *InformationCell = @"InformationCell";
     self.circularProgress.tintColor = [UIColor appTertiaryColor1];
     [self.circularProgressBar addSubview:self.circularProgress];
     
-    self.label3.text = [NSString stringWithFormat:@"%lu/%lu", completedScheduledTasks, allScheduledTasks];
+    self.label3.text = [NSString stringWithFormat:@"%lu/%lu", (unsigned long)completedScheduledTasks, (unsigned long)allScheduledTasks];
 
 }
